Move functions from `AndroidTemporalMapper` to `AndroidTimeUtils`
