+++ lib/src/main/kotlin/at/bitfire/synctools/mapping/calendar/handler/DurationHandler.kt
@@ -10,8 +10,8 @@ import android.content.Entity
 import android.provider.CalendarContract.Events
 import at.bitfire.ical4android.util.TimeApiExtensions.abs
 import at.bitfire.synctools.icalendar.plusAssign
-import at.bitfire.synctools.util.AndroidTemporalMapper.toZonedDateTime
 import at.bitfire.synctools.util.AndroidTimeUtils
+import at.bitfire.synctools.util.AndroidTimeUtils.toZonedDateTime
 import net.fortuna.ical4j.model.component.VEvent
 import net.fortuna.ical4j.model.property.DtEnd
 import java.time.Instant
